requirements.txt
Django==4.1
asgiref==3.5.2
dj-static==0.0.6
pytz==2020.4
sqlparse==0.4.1
static3==0.7.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
channels==2.4.0
Werkzeug==2.0.3
dynaconf==3.1.12
ipython==8.5.0
django-markdownx==4.0.2
django-filter==21.1
django-widget-tweaks==1.4.12